% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlychi.R
\name{get.lychi}
\alias{get.lychi}
\title{Compute Lychi hash for a molecule}
\usage{
get.lychi(smiles, keto_enol = FALSE)
}
\arguments{
\item{smiles}{A SMILES representation of the molecule}

\item{keto_enol}{If `TRUE` consider keto-enol tautomers, but can slow things down}
}
\value{
A string representing the Lychi hash
}
\description{
The Layered Chemical Identifier (Lychi) is a unique representation of a
molecule, similar in concept to the InChI key. In contrast to the InChI
key it is layered, meaning that the hash can be split into different layers
that allows one to compare molecules at different levels of resolution.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
